feat: add a main.py file
Added a main.py file that trains the agent that complements the notebook. I have also modified the train class, where now it takes an agent as an argument rather than creating an agent inside the class. This allows to pass any agent to the train method, for both training and evaluation without having to create one every time the class is created.